pyproject.toml
[tool.poetry]
name = "quiz_app"
version = "0.1.0"
description = "Quiz Management System using Postgresql and FastAPI Framework"
authors = ["Ramadhani , adyan.rendra@gmail.com"]

[tool.poetry.dependencies]
python = "^3.12.8"
fastapi = "^0.115.11"
uvicorn = "^0.34.0"
sqlalchemy = "^2.0.39"
psycopg2-binary = "^2.9.10"
Jinja2= "^3.1.6"
passlib = { extras = ["bcrypt"], version = "^1.7.4" }
python-dotenv = "^1.0.0"
pydantic = {version = "*", extras = ["email"]}
python-jose = "^3.3.0"
python-multipart = "^0.0.20"
fastapi-cache2 ="^0.2.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"